Restructure codebase
Goal here was to use a better project structure.
Got inspiration from Django
