--- ConsigliereDB.session.sql
@@ -1,1 +1,0 @@
-
