+++ consigliere/db/models/music_data.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
-from consigliere.data.db.database import Base
+from db.database import Base
 
 class MusicData(Base):
     __tablename__ = 'music_data'
